#37 added Exception handling and validation for messages

+++ backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/SimpleMessageService.java
@@ -15,6 +15,8 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 import at.ac.tuwien.sepr.groupphase.backend.service.UserService;
+import at.ac.tuwien.sepr.groupphase.backend.service.validators.MessageValidator;
+import org.hibernate.exception.ConstraintViolationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -30,25 +32,32 @@ public class SimpleMessageService implements MessageService {
     private static final Logger LOGGER =
         LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
     private final MessageRepository messageRepository;
-
     private final UserService userService;
+    private final MessageValidator validator;
 
     @Autowired
-    public SimpleMessageService(MessageRepository messageRepository, UserService userService) {
+    public SimpleMessageService(MessageRepository messageRepository, UserService userService, MessageValidator validator) {
         this.messageRepository = messageRepository;
         this.userService = userService;
+        this.validator = validator;
     }
 
     @Override
-    public List<ApplicationMessage> findAll() {
+    public List<ApplicationMessage> findAll() throws NotFoundException {
         LOGGER.debug("Find all messages");
         ApplicationUser user = userService.findApplicationUserByEmail(SecurityContextHolder.getContext().getAuthentication().getName());
-        return messageRepository.findAllByApplicationUserOrderByIsReadAscSentAtDesc(user);
+        List<ApplicationMessage> messages = messageRepository.findAllByApplicationUserOrderByIsReadAscSentAtDesc(user);
+        if (messages.isEmpty()) {
+            throw new NotFoundException(String.format("Could not find messages of user with email address %s",
+                user.getEmail()));
+        }
+        return messages;
     }
 
     @Override
-    public ApplicationMessage create(MessageCreateDto message) {
+    public ApplicationMessage create(MessageCreateDto message) throws ConstraintViolationException, ValidationException {
         LOGGER.debug("Publish new message {}", message);
+        validator.validateForCreate(message);
         ApplicationMessage applicationMessage = ApplicationMessage.ApplicationMessageBuilder.message()
             .withApplicationUser(userService.findApplicationUserById(message.getUserId()))
             .withGroupId(message.getGroupId())
@@ -61,14 +70,24 @@ public class SimpleMessageService implements MessageService {
     @Override
     public ApplicationMessage update(MessageDetailDto toUpdate) throws NotFoundException, ValidationException, ConflictException {
         LOGGER.debug("Update message {}", toUpdate);
-        ApplicationMessage myMessage = messageRepository.findById(toUpdate.getId()).orElseThrow();
-        myMessage.setIsRead(true);
+        ApplicationMessage myMessage = messageRepository.findById(toUpdate.getId()).orElse(null);
+        if (myMessage == null) {
+            throw new NotFoundException(String.format("Could not find message from group %s sent at %s",
+                toUpdate.getGroup().getName(), toUpdate.getSentAt()));
+        }
+        toUpdate.setIsRead(true);
+        validator.validateForUpdate(toUpdate);
+        myMessage.setIsRead(toUpdate.getIsRead());
         return messageRepository.save(myMessage);
     }
 
     @Override
-    public void delete(Long messageId) {
+    public void delete(Long messageId) throws NotFoundException {
         LOGGER.debug("Delete message with id {}", messageId);
+        if (!messageRepository.existsById(messageId)) {
+            throw new NotFoundException(String.format("Could not find message with id %s",
+                messageId));
+        }
         messageRepository.deleteById(messageId);
     }
